e2e tests for receive and send transactions flow
