+++ src/components/Alert.vue
@@ -1,5 +1,5 @@
 <template>
-  <div v-if="message" class="alert" :class="[type, { closed }]">
+  <div data-test="alert" v-if="message" class="alert" :class="[type, { closed }]">
     <div v-if="message" class="message">
       <p class="text">{{ message }}</p>
       <font-awesome-icon class="icon cross" icon="times" @click="close" />
@@ -11,7 +11,6 @@
 </template>
 
 <script>
-import { mapState } from 'vuex'
 
 export default {
   name: 'Alert',
@@ -42,13 +41,6 @@ export default {
       this.closed = true
       this.$emit('close')
     },
-  },
-  computed: {
-    ...mapState({
-      clearError: state => {
-        return state.wallet.clearError
-      },
-    }),
   },
 }
 </script>
