+++ src/components/Input.vue
@@ -1,6 +1,5 @@
 <template>
   <input
-    data-test="type-word-seed"
     v-model="inputValue"
     :class="`input ${getClass(type)}`"
     :placeholder="placeholder"
