+++ tests/e2e/specs/createWallet.js
@@ -1,0 +1,92 @@
+// https://docs.cypress.io/api/introduction/api.html
+import { createSelection } from '../utils'
+
+describe('Create Wallet', () => {
+  it('Create a wallet succesfully', () => {
+    // Cypress.config('baseUrl', 'http://localhost:8080')
+    cy.visit('/ftu')
+    cy.get('[data-test=create-wallet]').click()
+    cy.get('[data-test=do-it]').click()
+    cy.get('[data-test=new-seed-option]').click()
+    cy.get('[data-test=next-step]').click()
+    cy.get('[data-test=word-seed]')
+      .then(textarea => {
+        return createSelection(textarea, 0, 5)
+      })
+      .then(val => {
+        cy.get('[data-test=next-step]').click()
+        cy.get('input').type(val)
+      })
+    cy.get('[data-test=next-step]').click()
+    cy.get('[data-test=password-input]').type('password')
+    cy.get('[data-test=repeat-password]').type('password')
+    cy.get('[data-test=next-step]').click()
+    cy.get('[data-test=home]')
+  })
+  it('Show error when mnemonics don´t match', () => {
+    // Cypress.config('baseUrl', 'http://localhost:8080')
+    cy.visit('/ftu')
+    cy.get('[data-test=create-wallet]').click()
+    cy.get('[data-test=do-it]').click()
+    cy.get('[data-test=new-seed-option]').click()
+    cy.get('[data-test=next-step]').click()
+    cy.get('[data-test=word-seed]')
+      .then(textarea => {
+        return createSelection(textarea, 0, 5)
+      })
+      .then(val => {
+        cy.get('[data-test=next-step]').click()
+        cy.get('input').type(val + '1')
+      })
+    cy.get('[data-test=next-step]').click()
+    cy.get('[data-test=mnemonics-error-alert]').should('have.css', 'color', 'rgb(245, 108, 108)')
+  })
+  it('Show error when passwords don´t match', () => {
+    // Cypress.config('baseUrl', 'http://localhost:8080')
+    cy.visit('/ftu')
+    cy.get('[data-test=create-wallet]').click()
+    cy.get('[data-test=do-it]').click()
+    cy.get('[data-test=new-seed-option]').click()
+    cy.get('[data-test=next-step]').click()
+    cy.get('[data-test=word-seed]')
+      .then(textarea => {
+        return createSelection(textarea, 0, 5)
+      })
+      .then(val => {
+        cy.get('[data-test=next-step]').click()
+        cy.get('input').type(val)
+      })
+    cy.get('[data-test=next-step]').click()
+    cy.get('[data-test=password-input]').type('password')
+    cy.get('[data-test=repeat-password]').type('password1')
+    cy.get('[data-test=next-step]').click()
+    cy.get('[data-test=password-error-alert]').should('have.css', 'color', 'rgb(245, 108, 108)')
+  })
+  it('Show previous and next view when the button is clicked', () => {
+    // Cypress.config('baseUrl', 'http://localhost:8080')
+    cy.visit('/ftu')
+    cy.get('[data-test=create-wallet]').click()
+    cy.get('[data-test=do-it]').click()
+    cy.get('[data-test=new-seed-option]').click()
+    cy.get('[data-test=next-step]').click()
+    cy.get('[data-test=word-seed]')
+      .then(textarea => {
+        return createSelection(textarea, 0, 5)
+      })
+      .then(val => {
+        cy.get('[data-test=next-step]').click()
+        cy.get('input').type(val)
+      })
+    cy.get('[data-test=next-step]').click()
+    cy.get('[data-test=password-input]').type('password')
+    cy.get('[data-test=repeat-password]').type('password')
+    cy.get('[data-test=previous-step]').click()
+    cy.get('[data-test=header-4]')
+    cy.get('[data-test=previous-step]').click()
+    cy.get('[data-test=header-3]')
+    cy.get('[data-test=previous-step]').click()
+    cy.get('[data-test=header-2]')
+    cy.get('[data-test=previous-step]').click()
+    cy.get('[data-test=header-1]')
+  })
+})
