+++ tests/e2e/specs/sendTransaction.js
@@ -1,0 +1,31 @@
+describe('Create a complete Send transactions flow', () => {
+  it('Opens the modal for sending transactions and fill the form', () => {
+    cy.get('[data-test=send-btn]').click()
+    cy.get('[data-test=send-recipient-address]')
+      .click()
+      .type('x')
+    cy.get('input')
+      .first()
+      .click()
+      .type('x')
+    cy.get('input')
+      .last()
+      .click()
+      .type('1')
+    cy.get('[data-test=send-increase-amount]').click()
+    cy.get('[data-test=send-amount]').should('have.value', '2')
+    cy.get('[data-test=valueHigh]')
+    cy.get('[data-test=select-btn]').click()
+    cy.get('[data-test=option-Medium]').click()
+    cy.get('[data-test=valueMedium]')
+    cy.get('[data-test=sign-send-btn]').click()
+  })
+  it('Shows error when sending transactions', () => {
+    cy.get('[data-test=alert]')
+  })
+  it('Closes the modal', () => {
+    cy.get('.el-dialog__headerbtn')
+      .first()
+      .click({ force: true })
+  })
+})
