--- tests/e2e/specs/test.js
@@ -1,347 +1,0 @@
-// https://docs.cypress.io/api/introduction/api.html
-
-function createSelection(field, start, end) {
-  if (field.createTextRange) {
-    var selRange = field.createTextRange()
-    selRange.collapse(true)
-    selRange.moveStart('character', start)
-    selRange.moveEnd('character', end)
-    selRange.select()
-  } else if (field.setSelectionRange) {
-    field.setSelectionRange(start, end)
-  } else if (field.selectionStart) {
-    field.selectionStart = start
-    field.selectionEnd = end
-  }
-  field.focus()
-  return field[0].innerHTML
-}
-
-describe('Create Wallet', () => {
-  it('Create a wallet succesfully', () => {
-    // Cypress.config('baseUrl', 'http://localhost:8080')
-    cy.visit('/ftu')
-    cy.get('[data-test=create-wallet]').click()
-    cy.get('[data-test=do-it]').click()
-    cy.get('[data-test=new-seed-option]').click()
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=word-seed]')
-      .then(textarea => {
-        return createSelection(textarea, 0, 5)
-      })
-      .then(val => {
-        cy.get('[data-test=next-step]').click()
-        cy.get('[data-test=type-word-seed]').type(val)
-      })
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=password-input]').type('password')
-    cy.get('[data-test=repeat-password]').type('password')
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=home]')
-  })
-  it('Show error when mnemonics don´t match', () => {
-    // Cypress.config('baseUrl', 'http://localhost:8080')
-    cy.visit('/ftu')
-    cy.get('[data-test=create-wallet]').click()
-    cy.get('[data-test=do-it]').click()
-    cy.get('[data-test=new-seed-option]').click()
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=word-seed]')
-      .then(textarea => {
-        return createSelection(textarea, 0, 5)
-      })
-      .then(val => {
-        cy.get('[data-test=next-step]').click()
-        cy.get('[data-test=type-word-seed]').type(val + '1')
-      })
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=mnemonics-error-alert]').should('have.css', 'color', 'rgb(245, 108, 108)')
-  })
-  it('Show error when passwords don´t match', () => {
-    // Cypress.config('baseUrl', 'http://localhost:8080')
-    cy.visit('/ftu')
-    cy.get('[data-test=create-wallet]').click()
-    cy.get('[data-test=do-it]').click()
-    cy.get('[data-test=new-seed-option]').click()
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=word-seed]')
-      .then(textarea => {
-        return createSelection(textarea, 0, 5)
-      })
-      .then(val => {
-        cy.get('[data-test=next-step]').click()
-        cy.get('[data-test=type-word-seed]').type(val)
-      })
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=password-input]').type('password')
-    cy.get('[data-test=repeat-password]').type('password1')
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=password-error-alert]').should('have.css', 'color', 'rgb(245, 108, 108)')
-  })
-  it('Show previous and next view when the button is clicked', () => {
-    // Cypress.config('baseUrl', 'http://localhost:8080')
-    cy.visit('/ftu')
-    cy.get('[data-test=create-wallet]').click()
-    cy.get('[data-test=do-it]').click()
-    cy.get('[data-test=new-seed-option]').click()
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=word-seed]')
-      .then(textarea => {
-        return createSelection(textarea, 0, 5)
-      })
-      .then(val => {
-        cy.get('[data-test=next-step]').click()
-        cy.get('[data-test=type-word-seed]').type(val)
-      })
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=password-input]').type('password')
-    cy.get('[data-test=repeat-password]').type('password')
-    cy.get('[data-test=previous-step]').click()
-    cy.get('[data-test=header-4]')
-    cy.get('[data-test=previous-step]').click()
-    cy.get('[data-test=header-3]')
-    cy.get('[data-test=previous-step]').click()
-    cy.get('[data-test=header-2]')
-    cy.get('[data-test=previous-step]').click()
-    cy.get('[data-test=header-1]')
-  })
-})
-
-describe('Land correctly in every view accesible from the home page', () => {
-  it('Create a Wallet', () => {
-    cy.visit('/ftu')
-    cy.get('[data-test=create-wallet]').click()
-    cy.get('[data-test=do-it]').click()
-    cy.get('[data-test=new-seed-option]').click()
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=word-seed]')
-      .then(textarea => {
-        return createSelection(textarea, 0, 5)
-      })
-      .then(val => {
-        cy.get('[data-test=next-step]').click()
-        cy.get('[data-test=type-word-seed]').type(val)
-      })
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=password-input]').type('password')
-    cy.get('[data-test=repeat-password]').type('password')
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=home]')
-  })
-  it('Close session', () => {
-    cy.get('.el-dropdown-menu__item').click({ force: true })
-    cy.get('[data-test=welcome-back]')
-    // Be aware that when the password does not automatically display it should check if it is correct
-    cy.get('[data-test=welcome-back]')
-      .first()
-      .click()
-    cy.get('[data-test=password-input-access]').type('password')
-    cy.get('[data-test=unlock-wallet]').click()
-    cy.get('[data-test=home]')
-  })
-  it('Redirects to the correct social urls', () => {
-    cy.get('[data-test=to-community]')
-      .last()
-      .click()
-    cy.get('[data-test=community-page]')
-    cy.get('a').contains('Witnet Foundation Twitter account')
-    cy.get('a').contains('Witnet Community on Discord')
-    cy.get('a').contains('Witnet Foundation official blog on Medium')
-    cy.get('a').contains('Witnet Foundation contact email')
-    cy.get('a').contains('Witnet Community on Telegram')
-    cy.get('a').contains('Witnet GitHub repository')
-    cy.get('[data-test=logo-to-home]').click()
-    cy.get('[data-test=home]')
-  })
-  it('Redirects to Marketplace view', () => {
-    cy.get('[data-test=to-marketplace]').click()
-    cy.get('[data-test=marketplace]')
-    cy.get('[data-test=to-transactions]').click()
-    cy.get('[data-test=transactions]')
-  })
-})
-
-describe('Create a complete Receive transactions flow', () => {
-  it('Add Addresses correctly', () => {
-    cy.get('[data-test=receive-btn]').click()
-    cy.get('[data-test=receive-modal]')
-    cy.get('[data-test=last-address]')
-    cy.get('[data-test=copy-last-address]').click()
-  })
-  it('Copies last address generated and closes modal', () => {
-    cy.get('.el-dialog__headerbtn')
-      .last()
-      .click({ force: true })
-    cy.get('[data-test=receive-modal]').should(
-      'have.attr',
-      'style',
-      'z-index: 2001; display: none;'
-    )
-  })
-  it('Goes to pages and shows addresses when using pagination nav', () => {
-    cy.get('[data-test=address-1-0]')
-    cy.get('[data-test=address-1-1]')
-    cy.get('[data-test=address-1-2]')
-    cy.get('[data-test=address-1-3]')
-    cy.get('[data-test=address-1-4]')
-    cy.get('[data-test=page-1]').click()
-    cy.get('[data-test=address-1-0]')
-    cy.get('[data-test=address-1-1]')
-    cy.get('[data-test=address-1-2]')
-    cy.get('[data-test=address-1-3]')
-    cy.get('[data-test=address-1-4]')
-    cy.get('[data-test=page-2]').click()
-    cy.get('[data-test=address-2-0]')
-    cy.get('[data-test=address-2-1]')
-    cy.get('[data-test=address-2-2]')
-    cy.get('[data-test=address-2-3]')
-    cy.get('[data-test=address-2-4]')
-    cy.get('[data-test=page-3]').click()
-    cy.get('[data-test=address-3-0]')
-    cy.get('[data-test=address-3-1]')
-    cy.get('[data-test=address-3-2]')
-    cy.get('[data-test=address-3-3]')
-    cy.get('[data-test=address-3-4]')
-    cy.get('[data-test=page-4]').click()
-    cy.get('[data-test=address-4-0]')
-    cy.get('[data-test=address-4-1]')
-    cy.get('[data-test=address-4-2]')
-    cy.get('[data-test=address-4-3]')
-    cy.get('[data-test=address-4-4]')
-    cy.get('[data-test=page-5]').click()
-    cy.get('[data-test=address-5-0]')
-    cy.get('[data-test=address-5-1]')
-    cy.get('[data-test=address-5-2]')
-    cy.get('[data-test=address-5-3]')
-    cy.get('[data-test=address-5-4]')
-    cy.get('[data-test=page-6]').click()
-    cy.get('[data-test=address-6-0]')
-    cy.get('[data-test=paginate-to-left]').click()
-    cy.get('[data-test=address-5-0]')
-    cy.get('[data-test=address-5-1]')
-    cy.get('[data-test=address-5-2]')
-    cy.get('[data-test=address-5-3]')
-    cy.get('[data-test=address-5-4]')
-    cy.get('[data-test=paginate-to-left]').click()
-    cy.get('[data-test=address-4-0]')
-    cy.get('[data-test=address-4-1]')
-    cy.get('[data-test=address-4-2]')
-    cy.get('[data-test=address-4-3]')
-    cy.get('[data-test=address-4-4]')
-    cy.get('[data-test=paginate-to-left]').click()
-    cy.get('[data-test=address-3-0]')
-    cy.get('[data-test=address-3-1]')
-    cy.get('[data-test=address-3-2]')
-    cy.get('[data-test=address-3-3]')
-    cy.get('[data-test=address-3-4]')
-    cy.get('[data-test=paginate-to-left]').click()
-    cy.get('[data-test=address-2-0]')
-    cy.get('[data-test=address-2-1]')
-    cy.get('[data-test=address-2-2]')
-    cy.get('[data-test=address-2-3]')
-    cy.get('[data-test=address-2-4]')
-    cy.get('[data-test=paginate-to-left]').click()
-    cy.get('[data-test=address-1-0]')
-    cy.get('[data-test=address-1-1]')
-    cy.get('[data-test=address-1-2]')
-    cy.get('[data-test=address-1-3]')
-    cy.get('[data-test=address-1-4]')
-    cy.get('[data-test=paginate-to-right]').click()
-    cy.get('[data-test=address-2-0]')
-    cy.get('[data-test=address-2-1]')
-    cy.get('[data-test=address-2-2]')
-    cy.get('[data-test=address-2-3]')
-    cy.get('[data-test=address-2-4]')
-    cy.get('[data-test=paginate-to-right]').click()
-    cy.get('[data-test=address-3-0]')
-    cy.get('[data-test=address-3-1]')
-    cy.get('[data-test=address-3-2]')
-    cy.get('[data-test=address-3-3]')
-    cy.get('[data-test=address-3-4]')
-    cy.get('[data-test=paginate-to-right]').click()
-    cy.get('[data-test=address-4-0]')
-    cy.get('[data-test=address-4-1]')
-    cy.get('[data-test=address-4-2]')
-    cy.get('[data-test=address-4-3]')
-    cy.get('[data-test=address-4-4]')
-    cy.get('[data-test=paginate-to-right]').click()
-    cy.get('[data-test=address-5-0]')
-    cy.get('[data-test=address-5-1]')
-    cy.get('[data-test=address-5-2]')
-    cy.get('[data-test=address-5-3]')
-    cy.get('[data-test=address-5-4]')
-    cy.get('[data-test=paginate-to-right]').click()
-    cy.get('[data-test=address-6-0]')
-    cy.get('[data-test=paginate-to-right]').should('not.exist')
-    cy.get('[data-test=page-1]').click()
-    cy.get('[data-test=paginate-to-left]').should('not.exist')
-  })
-  it('Create a Wallet', () => {
-    cy.visit('/ftu')
-    cy.get('[data-test=create-wallet]').click()
-    cy.get('[data-test=do-it]').click()
-    cy.get('[data-test=new-seed-option]').click()
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=word-seed]')
-      .then(textarea => {
-        return createSelection(textarea, 0, 5)
-      })
-      .then(val => {
-        cy.get('[data-test=next-step]').click()
-        cy.get('[data-test=type-word-seed]').type(val)
-      })
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=password-input]').type('password')
-    cy.get('[data-test=repeat-password]').type('password')
-    cy.get('[data-test=next-step]').click()
-    cy.get('[data-test=home]')
-  })
-  it('shows only one page whith less than 5 addresses generated', () => {
-    cy.get('[data-test=receive-btn]').click()
-    cy.get('[data-test=receive-modal]')
-    cy.get('[data-test=last-address]')
-    cy.get('[data-test=copy-last-address]').click()
-    cy.get('.el-dialog__headerbtn')
-      .last()
-      .click({ force: true })
-    cy.get('[data-test=page-1]')
-      .first()
-      .click()
-    cy.get('[data-test=page-2]').should('not.exist')
-    cy.get('[data-test=receive-btn]').click()
-    cy.get('[data-test=receive-modal]')
-    cy.get('[data-test=last-address]')
-    cy.get('[data-test=copy-last-address]').click()
-    cy.get('.el-dialog__headerbtn')
-      .last()
-      .click({ force: true })
-    cy.get('[data-test=receive-btn]').click()
-    cy.get('[data-test=receive-modal]')
-    cy.get('[data-test=last-address]')
-    cy.get('[data-test=copy-last-address]').click()
-    cy.get('.el-dialog__headerbtn')
-      .last()
-      .click({ force: true })
-    cy.get('[data-test=receive-btn]').click()
-    cy.get('[data-test=receive-modal]')
-    cy.get('[data-test=last-address]')
-    cy.get('[data-test=copy-last-address]').click()
-    cy.get('.el-dialog__headerbtn')
-      .last()
-      .click({ force: true })
-    cy.get('[data-test=receive-btn]').click()
-    cy.get('[data-test=receive-modal]')
-    cy.get('[data-test=last-address]')
-    cy.get('[data-test=copy-last-address]').click()
-    cy.get('.el-dialog__headerbtn')
-      .last()
-      .click({ force: true })
-    cy.get('[data-test=receive-btn]').click()
-    cy.get('[data-test=receive-modal]')
-    cy.get('[data-test=last-address]')
-    cy.get('[data-test=copy-last-address]').click()
-    cy.get('.el-dialog__headerbtn')
-      .last()
-      .click({ force: true })
-  })
-})
